refactor: dispatch and pickup split, investigating shared-code opportunities
